Change hcode to long to match the libcurl documentation

+++ src/eb.h
@@ -211,7 +211,7 @@ extern bool inInput;		/* reading line from standard in */
 extern int fileSize;		/* when reading/writing files */
 extern int maxFileSize;		/* max size of an editable file */
 extern int mssock;		/* mail server socket */
-extern int hcode;		/* http code, like 404 file not found */
+extern long hcode;		/* http code, like 404 file not found */
 extern char herror[];		/* html error */
 extern char errorMsg[];		/* generated error message */
 extern char serverLine[];	/* lines to and from the mail server */
